Disable language selection & auto-discovery
