+++ web/templates/base.html
@@ -147,6 +147,7 @@
             <i class="fa-brands fa-github"></i>
             {{ .T "Github" }}
           </a>
+          <!-- XXX: Disabled Oct 2024 due to very poor coverage
           <div
             class="ui scrolling dropdown"
             style="margin-left: 5px;"
@@ -176,7 +177,7 @@
                 </div>
               {{ end }}
             </div>
-          </div>
+          </div> -->
         </div>
       </center>
